@sandersn 2020-07-13
Just to confirm, this fix only helps if you explicitly annotate variables with `CssFunctionReturnType`, right?