@ardeois 2020-07-14
> Oh, never mind, the slow part is the emit, not the type checking. Exporting the type makes it available for use when writing the .d.ts.
Yes exactly, thanks for merging @sandersn